Add files via upload
I've got the basics down for the ability score page. It's pretty bare bones, but it has all the essential parts. All that needs to be finished is making it look nice with better CSS that matches the left sidenav, and writing some static text for the rightnav. You can play around with it to see if you can catch any bugs that I missed. Let me know and I'll get on it. (My code is a mess, I doubt you'd be able to fix any bugs you find)